requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
PyJWT==2.1.0
SQLAlchemy==1.4.20
Flask-SQLAlchemy==2.5.1